Requests to API using JWT + Clean, fixes and identation
