+++ guardiansFrontend/package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "axios": "^0.18.0",
     "bootstrap": "^4.1.3",
+    "cors": "^2.8.5",
     "immutability-helper": "^2.9.0",
     "jquery": "^3.3.1",
     "react": "^16.7.0",
@@ -20,7 +21,7 @@
     "react-sidebar": "^3.0.2"
   },
   "scripts": {
-    "start": "set PORT=3001 && xreact-scripts start",
+    "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
