+++ guardiansFrontend/src/Components/app/teams/TeamsComponent.js
@@ -1,5 +1,5 @@
 import React from 'react';
-//import './src/Components/css/teamsComp.css';
+import './css/teamsComp.css';
 import {fetchProfiles} from './fetchData';
 
 class TeamsComponent extends React.Component {
@@ -31,18 +31,6 @@ class TeamsComponent extends React.Component {
               });
             }
           )
-      }
-
-      componentDidMount() {
-        fetchProfiles("teams")
-        .then(r =>{
-          return r.json();
-        })
-        .then(data => {
-          console.log(data);
-          this.setState({participants: data, isLoaded:true})
-        })
-        .catch(error => this.setState({error:"CRASHOU", isLoaded:true}))
       }
     render(){
         const { isLoaded, items } = this.state;
