+++ guardiansFrontend/src/Components/app/teams/css/teamsComp.css
@@ -1,4 +1,3 @@
-
 .teamsContainer{
   display: flex;
   height: 600px;
