+++ guardiansFrontend/src/Components/app/teams/fetchData.js
@@ -1,14 +1,26 @@
+export function getToken() {
+  // Retrieves the user token from localStorage
+  console.log(localStorage.getItem('Authorization'))
+  return localStorage.getItem('Authorization')
+}
 
 export function fetchProfiles(url) {
-    return fetch('http://localhost:3000/'+url)   
+    return fetch('https://guardianshackatum.herokuapp.com/api/v1/'+url ,{ 
+      headers: {
+        'Accept': 'application/json',
+        'Content-Type': 'application/json',
+        'Authorization' : 'Bearer ' +  getToken()
+      }
+    })   
 }
 
 export function  getTeamFromApiAsync(data) {
-  return fetch('http://localhost:3000/teams', {
+  return fetch('https://guardianshackatum.herokuapp.com/api/v1/teams', {
     method: 'post',
     headers: {
       'Accept': 'application/json',
       'Content-Type': 'application/json',
+      'Authorization' : 'Bearer ' +  getToken()
     },
     body: JSON.stringify({
       name: data.nameTeam,
@@ -30,11 +42,12 @@ export function  updateTeamFromApiAsync(data) {
   var pageURL = window.location.href;
   var lastURLSegment = pageURL.substr(pageURL.lastIndexOf('/') + 1);
   console.log(lastURLSegment)
-  return fetch('http://localhost:3000/teams/'+lastURLSegment, {
+  return fetch('https://guardianshackatum.herokuapp.com/api/v1/teams/'+lastURLSegment, {
     method: 'put',
     headers: {
       'Accept': 'application/json',
       'Content-Type': 'application/json',
+      'Authorization' : 'Bearer ' +  getToken()
     },
     body: JSON.stringify({
       name: data.nameTeam,
@@ -54,11 +67,12 @@ export function  updateTeamFromApiAsync(data) {
 
 export function  DeleteTeamFromApiAsync(id) {
 
-  return fetch('http://localhost:3000/teams/'+id, {
+  return fetch('https://guardianshackatum.herokuapp.com/api/v1/teams/'+id, {
     method: 'delete',
     headers: {
       'Accept': 'application/json',
       'Content-Type': 'application/json',
+      'Authorization' : 'Bearer ' +  getToken()
     },
   }).then((response) => response.json())
     .then((responseJson) => {
